Moved the report related functionality to its own class.
