+++ includes/Modules/AdSense.php
@@ -32,9 +32,12 @@ use Google\Site_Kit\Core\Modules\Module_With_Service_Entity;
 use Google\Site_Kit\Core\REST_API\Data_Request;
 use Google\Site_Kit\Core\Tags\Guards\Tag_Environment_Type_Guard;
 use Google\Site_Kit\Core\Tags\Guards\Tag_Verify_Guard;
+use Google\Site_Kit\Core\Util\Date;
 use Google\Site_Kit\Core\Util\Debug_Data;
 use Google\Site_Kit\Core\Util\Feature_Flags;
 use Google\Site_Kit\Core\Util\Method_Proxy_Trait;
+use Google\Site_Kit\Core\Util\Sort;
+use Google\Site_Kit\Core\Util\URL;
 use Google\Site_Kit\Modules\AdSense\AMP_Tag;
 use Google\Site_Kit\Modules\AdSense\Settings;
 use Google\Site_Kit\Modules\AdSense\Tag_Guard;
@@ -45,8 +48,6 @@ use Google\Site_Kit_Dependencies\Google\Service\Adsense as Google_Service_Adsens
 use Google\Site_Kit_Dependencies\Google\Service\Adsense\Alert as Google_Service_Adsense_Alert;
 use Google\Site_Kit_Dependencies\Psr\Http\Message\RequestInterface;
 use Exception;
-use Google\Site_Kit\Core\Util\Sort;
-use Google\Site_Kit\Core\Util\URL;
 use WP_Error;
 
 /**
@@ -528,7 +529,7 @@ final class AdSense extends Module
 			case 'last-14-days':
 			case 'last-28-days':
 			case 'last-90-days':
-				return $this->parse_date_range( $date_range );
+				return Date::parse_date_range( $date_range );
 		}
 
 		return new WP_Error( 'invalid_date_range', __( 'Invalid date range.', 'google-site-kit' ) );
@@ -611,7 +612,7 @@ final class AdSense extends Module
 				function ( $hostname ) {
 					return 'DOMAIN_NAME==' . $hostname;
 				},
-				$this->permute_site_hosts( $site_hostname )
+				URL::permute_site_hosts( $site_hostname )
 			)
 		);
