+++ includes/Modules/Site_Verification.php
@@ -23,6 +23,7 @@ use Google\Site_Kit\Core\REST_API\Data_Request;
 use Google\Site_Kit\Core\Util\Exit_Handler;
 use Google\Site_Kit\Core\Util\Google_URL_Matcher_Trait;
 use Google\Site_Kit\Core\Util\Method_Proxy_Trait;
+use Google\Site_Kit\Core\Util\URL;
 use Google\Site_Kit_Dependencies\Google\Service\Exception as Google_Service_Exception;
 use Google\Site_Kit_Dependencies\Google\Service\SiteVerification as Google_Service_SiteVerification;
 use Google\Site_Kit_Dependencies\Google\Service\SiteVerification\SiteVerificationWebResourceGettokenRequest as Google_Service_SiteVerification_SiteVerificationWebResourceGettokenRequest;
@@ -203,7 +204,7 @@ final class Site_Verification extends Module implements Module_With_Scopes {
 						$restore_defer = $this->with_client_defer( false );
 						$errors        = new WP_Error();
 
-						foreach ( $this->permute_site_url( $data['siteURL'] ) as $url ) {
+						foreach ( URL::permute_site_url( $data['siteURL'] ) as $url ) {
 							$site = new Google_Service_SiteVerification_SiteVerificationWebResourceResourceSite();
 							$site->setType( 'SITE' );
 							$site->setIdentifier( $url );
