+++ tests/phpunit/includes/Core/Modules/FakeModule.php
@@ -247,27 +247,6 @@ class FakeModule extends Module
 		return parent::exception_to_error( $e, $datapoint );
 	}
 
-
-	/**
-	 * Parses a date range string into a start date and an end date.
-	 *
-	 * @since 1.0.0
-	 *
-	 * @param string $range         Date range string. Either 'last-7-days', 'last-14-days', 'last-90-days', or
-	 *                              'last-28-days' (default).
-	 * @param string $multiplier    Optional. How many times the date range to get. This value can be specified if the
-	 *                              range should be request multiple times back. Default 1.
-	 * @param int    $offset        Days the range should be offset by. Default 1. Used by Search Console where
-	 *                              data is delayed by two days.
-	 * @param bool   $previous      Whether to select the previous period. Default false.
-	 *
-	 * @return array List with two elements, the first with the start date and the second with the end date, both as
-	 *               'Y-m-d'.
-	 */
-	public function parse_date_range( $range, $multiplier = 1, $offset = 1, $previous = false ) { // phpcs:ignore Generic.CodeAnalysis.UselessOverridingMethod
-		return parent::parse_date_range( $range, $multiplier, $offset, $previous );
-	}
-
 	/**
 	 * Gets the module's owner ID.
 	 *
